split observable/registry.rb
